pipeline/requirements.txt
# Core dependencies
pandas>=2.0.0
pyarrow>=14.0.0
polars>=0.20.0
numpy>=1.24.0

# Data processing
geopandas>=0.14.0
shapely>=2.0.0
h3>=3.7.0

# ML
scikit-learn>=1.3.0
xgboost>=2.0.0

# Azure upload
azure-storage-blob>=12.19.0

# CLI
click>=8.1.0
tqdm>=4.66.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
ruff>=0.1.0
